Выполнение задания 32
