+++ app/build.gradle
@@ -53,4 +53,6 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
+    implementation "androidx.lifecycle:lifecycle-extensions:2.2.0"
+    implementation("androidx.fragment:fragment-ktx:1.6.1")
 }
