+++ app/src/main/java/com/example/android_intern/MainActivity.kt
@@ -3,14 +3,11 @@ package com.example.android_intern
 
 import android.os.Bundle
 import android.util.Log
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.android_intern.databinding.ActivityMainBinding
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
 
 const val BASE_URL = "https://api.openweathermap.org/data/2.5/"
 const val APP_ID = "eeba719e0ea1ed0d70d6ea433307695e"
@@ -21,45 +18,17 @@ const val TAG = "Debug"
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
     private val adapter = WeatherAdapter()
-    private lateinit var apiService: WeatherApi
+    private val viewModel by viewModels<ForecastViewModel>()
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
-        initApiService()
         initWeatherRecyclerView()
-        loadWeather()
     }
 
-    private fun loadWeather() {
-        apiService.getCurrentForecastData(cityId = CITY_ID, appId = APP_ID, units = UNITS)
-            .enqueue(object : Callback<ForecastResponse> {
-                override fun onResponse(
-                    call: Call<ForecastResponse>,
-                    response: Response<ForecastResponse>
-                ) {
-                    if (response.isSuccessful) {
-                        val forecastGetList = response.body()!!
-                        runOnUiThread {
-                            adapter.submitList(forecastGetList.list)
-                        }
-                    } else {
-                        Log.e(TAG, binding.root.context.getString(R.string.error_connect))
-                    }
-                }
+    override fun onStart() {
+        super.onStart()
 
-                override fun onFailure(call: Call<ForecastResponse>, t: Throwable) {
-                    Log.e(TAG, t.message,t)
-                }
-            })
-    }
-
-    private fun initApiService() {
-        val retrofit = Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-        apiService = retrofit.create(WeatherApi::class.java)
     }
 
     private fun initWeatherRecyclerView() {
